#34 added a lateral bar with scroll
That was not easy as spected, so the scrollbar also has not appeared
once again. After searching for loooooooong time, i found nothing, not
even 1 link talking about the same problem as me, so i started to debug
my code with no clue of logic related to the proble. Until I've put the
same xml code in another activity, and worked!! but not in the
MainActivity where I need. So that was even more ilogical, and further
of any clue. Al last i've found out that when I set the main layout to
screen size, the fucking scroll worked(normally is set at 10000, the
size of the graph board). So the logic was the scrollView inner Layout
is too short so that the scrollView detect that needs scrollbars, so the
solution was set a higher size than the main layout so the scrollview
can detect the overflow