minor bug fix at beautify
With the capability of the nodes to have more fathers, according the
positioning of nodes with odd sons, nodes that only have 1 son, and this
is shared with another node, where positioned one over the other, and
that was not correct. now depends on if the son is already visited or
not(not visited behave as odd number of sons, visited behave as pair
number of son).